backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
fastapi==0.116.1
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
ollama==0.5.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.2
qdrant-client==1.15.1
regex==2025.9.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
tiktoken==0.11.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
